feat: バックエンドで放置していたシステムを実装

+++ src/main/java/io/github/bucheapp/roost/services/ChatServiceImpl.java
@@ -1,6 +1,8 @@
 package io.github.bucheapp.roost.services;
 
+import java.io.IOException;
 import java.time.LocalDateTime;
+import java.util.UUID;
 
 import jakarta.transaction.Transactional;
 
@@ -12,9 +14,11 @@ import org.springframework.web.multipart.MultipartFile;
 import org.springframework.web.server.ResponseStatusException;
 
 import io.github.bucheapp.roost.dto.request.ChatRequest;
+import io.github.bucheapp.roost.models.ApprovalChat;
 import io.github.bucheapp.roost.models.Chat;
 import io.github.bucheapp.roost.models.ChatType;
 import io.github.bucheapp.roost.models.Community;
+import io.github.bucheapp.roost.models.CommunityState;
 import io.github.bucheapp.roost.models.Room;
 import io.github.bucheapp.roost.models.TextChat;
 import io.github.bucheapp.roost.models.User;
@@ -37,6 +41,9 @@ public class ChatServiceImpl implements ChatService {
 	private RoomRepository roomRepository;
 	
 	@Autowired
+	private FileService fileService;
+	
+	@Autowired
 	private AuthContext authContext;
 	
 	@Autowired
@@ -56,7 +63,7 @@ public class ChatServiceImpl implements ChatService {
 
 	@Override
 	@Transactional
-	public Chat createChat(long publicId, ChatRequest req) {
+	public Chat createChat(long publicId, ChatRequest req) throws IOException {
 		long userId = authContext.getCurrentUserId();
 		ChatType type = req.getType();
 		
@@ -67,16 +74,17 @@ public class ChatServiceImpl implements ChatService {
 				.orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, messageUtil.get("user.notfound")));
 		
 		Snowflake snowflake = new Snowflake(workerIdProvider.getWorkerId(), datacenterId);
-		LocalDateTime now = LocalDateTime.now();
 		
 		Chat chat = null;
 		
 		if(type == ChatType.TEXT) {
 			MultipartFile file = req.getFile();
+			UUID uuid = null;
 			if(file != null) {
-				checkByte(req.getFile());
-				//TODO: ファイルを作成するロジックを作成
+				fileService.checkByte(file, 10 * 1024 * 1024);
+				uuid = fileService.createImage("image",file);
 			}
+			
 			chat = new TextChat(
 					snowflake.nextId(),
 					LocalDateTime.now(),
@@ -84,21 +92,25 @@ public class ChatServiceImpl implements ChatService {
 					room,
 					req
 					);
+			
+			((TextChat) chat).setMediaContentUUID(uuid);
+		} else if(type == ChatType.APPROVAL) {
+			chat = new ApprovalChat(
+					snowflake.nextId(),
+					LocalDateTime.now(),
+					user,
+					room
+					);
 		} else {
 			throw new ResponseStatusException(HttpStatus.BAD_REQUEST, messageUtil.get("unknown.chattype"));
 		}
-		
-		chat.setPublicId(snowflake.nextId());
-		chat.setCreatedAt(now);
-		chat.setRoom(room);
-		chat.setCreator(user);
 		
 		return chatRepository.save(chat);
 	}
 
 	@Override
 	@Transactional
-	public Chat updateChat(long publicId, ChatRequest req) {
+	public Chat updateChat(long publicId, ChatRequest req) throws IOException {
 		ChatType type = req.getType();
 		long userId = authContext.getCurrentUserId();
 		
@@ -118,8 +130,10 @@ public class ChatServiceImpl implements ChatService {
 			TextChat textChat = (TextChat) chat;
 			textChat.setContent(req.getContent());
 			if(file != null) {
-				checkByte(req.getFile());
-				//TODO: ファイルを作成するロジックを作成
+				fileService.checkByte(file, 10 * 1024 * 1024);
+				fileService.deleteImage("image/" + textChat.getMediaContentUUID() + ".jpg");
+				UUID uuid = fileService.createImage("image",file);
+				textChat.setMediaContentUUID(uuid);
 			}
 		} else {
 			throw new ResponseStatusException(HttpStatus.BAD_REQUEST, messageUtil.get("unknown.chattype"));
@@ -129,7 +143,10 @@ public class ChatServiceImpl implements ChatService {
 		
 		Room room = chat.getRoom();
 		Community community = room.getCommunity();
-		community.checkStateActive();
+		
+		if(community.getState() != CommunityState.ACTIVE) {
+			throw new ResponseStatusException(HttpStatus.FORBIDDEN, messageUtil.get("community.notactive"));
+		}
 		
 		return chatRepository.save(chat);
 	}
@@ -141,6 +158,10 @@ public class ChatServiceImpl implements ChatService {
 		
 		Chat chat = chatRepository.findByPublicId(publicId)
 				.orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, messageUtil.get("chat.notfound")));
+		
+		if(chat instanceof TextChat textChat) {
+			fileService.deleteImage("image/" + textChat.getMediaContentUUID() + ".jpg");
+		}
 		
 		Room room = chat.getRoom();
 		Community community = room.getCommunity();
@@ -154,19 +175,10 @@ public class ChatServiceImpl implements ChatService {
 			throw new ResponseStatusException(HttpStatus.FORBIDDEN, messageUtil.get("cannot.delete.chat"));
 		}
 		
-		community.checkStateActive();
+		if(community.getState() != CommunityState.ACTIVE) {
+			throw new ResponseStatusException(HttpStatus.FORBIDDEN, messageUtil.get("community.notactive"));
+		}
 		
 		chatRepository.delete(chat);
-	}
-
-	@Override
-	public void checkByte(MultipartFile file) {
-		if(file.isEmpty()) {
-			throw new ResponseStatusException(HttpStatus.BAD_REQUEST, messageUtil.get("file.isempty"));
-		}
-		
-		if (file.getSize() > 10 * 1024 * 1024) {
-			throw new ResponseStatusException(HttpStatus.BAD_REQUEST, messageUtil.get("file.size.exceeds"));
-		}
 	}
 }
